+++ CHANGE_LOG.md
@@ -5,6 +5,33 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project/module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## V0.6.0 - 27.01.2025
+
+### Added
+ - Support for CLI over SCP
+ - Support for enter bootloader over SCP command
+
+---
+## V0.5.0 - 16.10.2024
+
+### Added
+ - Added support for USB DFU via configuration switch
+ - Added config switch for FW image validation
+
+### Changed
+ - Support for different application header in BOOT menu, compatible with Revision module V2.0.0
+ - New bootloader interface implemented, with different "prepare" command
+
+---
+## V0.4.1 - 25.06.2024
+
+### Notice
+ This version has new CLI interface format, compatible with CLI V2.0.0 on
+
+### Changed
+ - Parameter info command addopted to CLI V2.0.0
+ 
+---
 ## V0.4.0 - 12.01.2024
 
 ### Notice